+++ client/shreds-ts/package.json
@@ -21,21 +21,22 @@
     "./*"
   ],
   "scripts": {
-    "lint": "eslint --ext .ts,.js --fix .",
     "format": "prettier --write .",
-    "build": "tsc -p .",
-    "dev": "tsc -p . && node dist/index.js",
+    "build": "tsup",
+    "dev": "tsx src/index.ts",
     "deploy:prod": "wrangler deploy",
     "update:packages": "pnpm update --latest"
   },
   "dependencies": {
-    "@validators-dao/solana-stream-sdk": "workspace:*",
     "@solana/web3.js": "1.98.2",
+    "@validators-dao/solana-stream-sdk": "workspace:*",
     "dotenv": "16.5.0"
   },
   "devDependencies": {
     "@types/node": "22.15.27",
     "prettier": "3.5.3",
+    "tsup": "8.5.0",
+    "tsx": "4.19.4",
     "typescript": "5.8.3"
   }
 }
