+++ temp-release/shreds-ts/package.json
@@ -21,10 +21,9 @@
     "./*"
   ],
   "scripts": {
-    "lint": "eslint --ext .ts,.js --fix .",
     "format": "prettier --write .",
-    "build": "tsc -p .",
-    "dev": "tsc -p . && node dist/index.js",
+    "build": "tsup",
+    "dev": "tsx src/index.ts",
     "deploy:prod": "wrangler deploy",
     "update:packages": "pnpm update --latest"
   },
@@ -36,6 +35,8 @@
   "devDependencies": {
     "@types/node": "22.15.27",
     "prettier": "3.5.3",
+    "tsup": "8.5.0",
+    "tsx": "4.19.4",
     "typescript": "5.8.3"
   }
 }
